backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.9,<0.116.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "passlib[bcrypt] (==1.7.4)",
    "bcrypt (==4.0.1)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "openai (>=1.77.0,<2.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "boto3 (>=1.37.37,<2.0.0)",
    "chromadb (>=1.0.7,<2.0.0)",
    "openai-agents (>=0.0.14,<0.0.15)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
